requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
aniso8601==8.0.0
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
azure-core==1.1.1
azure-storage-blob==12.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
flask-restplus==0.13.0
flask-swagger-ui==3.20.9
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.3.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.0.2
msrest==0.6.10
multidict==4.7.4
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyrsistent==0.15.6
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
yarl==1.4.2
zipp==0.6.0
